pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepwebharvester"
version = "2.0.0"
description = "Advanced OSINT dark web intelligence gathering tool via Tor"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "Tech Enthusiast" }]
keywords = ["osint", "tor", "darkweb", "cybersecurity", "scraper", "intelligence"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Security",
]
dependencies = [
    "requests>=2.31.0",
    "PySocks>=1.7.1",
    "stem>=1.8.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
]

[project.scripts]
deepwebharvester = "deepwebharvester.cli:main"

[project.urls]
Homepage = "https://github.com/servais1983/deepwebharvester"
Issues   = "https://github.com/servais1983/deepwebharvester/issues"

# ── Tool configuration ────────────────────────────────────────────────────────

[tool.setuptools.packages.find]
where = ["."]
include = ["deepwebharvester*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = [
    "--cov=deepwebharvester",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "-v",
]

[tool.coverage.run]
source   = ["deepwebharvester"]
omit     = ["*/tests/*"]

[tool.coverage.report]
show_missing    = true
skip_covered    = false
fail_under      = 80
